[testing] - Changes in routing | header, base and add categories section
