+++ pyproject.toml
@@ -9,6 +9,7 @@ dependencies = [
     "jinja2>=3.1.6",
     "pydantic>=2.11.7",
     "python-dotenv>=1.1.1",
+    "typing>=3.10.0.0",
 ]
 
 [dependency-groups]
