+++ templates/base.jinja
@@ -44,6 +44,9 @@
             {% block content %}
             {% endblock content %}
 
+            {% block layout %}
+            {% endblock layout %}
+
             {% block footer %}
                 {% include "partials/footer.jinja" %}
             {% endblock footer %}
