+++ templates/components/btn_link.jinja
@@ -1,5 +1,5 @@
 {% macro btn_link(text, url, hxTarget="#content") %}
-    <a class="btn-link"
+    <a class="inline-block text-black hover:text-lime-500 font-medium"
        hx-target="{{ hxTarget }}"
        hx-push-url="true"
        hx-get="{{ url }}">{{ text }}</a>
