+++ templates/pages/categories.jinja
@@ -1,0 +1,17 @@
+{% extends 'base.jinja' %}
+
+{% block layout %}
+    <h2 class="text-2xl font-semibold mb-4">Dashboard Layout</h2>
+
+    <div class="grid grid-cols-4 grid-rows-4 gap-4">
+        <div>01</div>
+        <div>02</div>
+        <div>03</div>
+        <div>04</div>
+        <div>05</div>
+        <div>06</div>
+        <div>07</div>
+        <div>08</div>
+        <div>09</div>
+    </div>
+{% endblock layout %}
