+++ templates/partials/header.jinja
@@ -1,10 +1,35 @@
 {% from 'components/btn_link.jinja' import btn_link %}
-<header class="fixed z-50 m-4 rounded-xl border min-w-lg bg-background/70 backdrop-blur">
+{# <header class="fixed z-50 m-4 rounded-xl border min-w-lg bg-background/70 backdrop-blur">
     <div class="flex flex-row justify-between items-center py-2 px-4 h-14">
-        <a class="text-xl font-medium hover:underline underline-offset-4">Caminatas</a>
+        <a class="text-xl font-medium hover:underline underline-offset-4">Company Name</a>
         <div class="flex flex-row justify-center">
-            {{ btn_link(url="/", text="home") }}
-            {{ btn_link(url="/about", text="about") }}
+            {{ btn_link(url="/", text="Home") }}
+            {{ btn_link(url="/about", text="About") }}
+            {{ btn_link(url="/categories", text="Categories") }}
         </div>
     </div>
+</header> #}
+<header class="fixed z-50 m-4 border w-full">
+    <section class="relative bg-teal-900">
+        <nav class="py-6">
+            <div class="container mx-auto px-4>
+                <div class="relative flex items-center justify-between">
+                    <a href="#" class="inline-block"> Company Name </a>
+
+                    <ul class="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 hidden md:flex">
+                        <li class="mr-8">
+                            {{ btn_link(url="/", text="Home") }}
+                        </li>
+                        <li class="mr-8">
+                            {{ btn_link(url="/about", text="About") }}
+                        </li>
+                        <li class="mr-8">
+                            {{ btn_link(url="/categories", text="Categories") }}
+                        </li>
+                    </ul>
+
+                </div>
+            </div>
+        </nav>
+    </section>
 </header>
