+++ test.py
@@ -1,0 +1,4 @@
+test = (1, 2, [3, 4])
+test[2].append(5)
+
+print(test)
